move functions to helpers.py to avoid circular import issue
