+++ tests/test_naki_and_actions.py
@@ -591,7 +591,6 @@ class TestRemainsAreSets(unittest.TestCase):
             check_remains_are_sets(remain_tiles, huro_count), False)
 
 
-# TODO: move to test utils
 class TestConsistsJantouAndSets(unittest.TestCase):
 
     def setUp(self):
