+++ tests/test_utils.py
@@ -1,7 +1,7 @@
 import unittest
 
-from mahjong.components import Suit, Tile
-from mahjong.utils import get_values, get_name, roundup, is_yaochuu
+from mahjong.components import Suit
+from mahjong.utils import get_values, get_name, roundup
 
 
 class TestGetValue(unittest.TestCase):
@@ -23,19 +23,4 @@ class TestRoundup(unittest.TestCase):
         self.assertEqual(roundup(1), 100)
 
 
-class TestIsYaochuu(unittest.TestCase):
 
-    def setUp(self):
-        self.yaochuu_tile_1 = Tile(Suit.JIHAI.value, 1)
-        self.yaochuu_tile_2 = Tile(Suit.MANZU.value, 9)
-        self.not_yaochuu_tile = Tile(Suit.SOUZU.value, 5)
-
-    def test_is_yaochuu(self):
-        test_1 = is_yaochuu(self.yaochuu_tile_1.suit, self.yaochuu_tile_1.rank)
-        test_2 = is_yaochuu(self.yaochuu_tile_2.suit, self.yaochuu_tile_2.rank)
-        test_3 = is_yaochuu(self.not_yaochuu_tile.suit,
-                            self.not_yaochuu_tile.rank)
-
-        self.assertEqual(test_1, True)
-        self.assertEqual(test_2, True)
-        self.assertEqual(test_3, False)
